requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.31.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.5
pydantic==2.5.3
pydantic-settings==2.1.0
openai==1.12.0
httpx==0.26.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0